re-org file's location based on review
