+++ src/components/settings/factory.js
