--- src/containers/common/CollapsibleContainer.css
@@ -1,9 +1,0 @@
-@import '../../styles/colors.css';
-
-.collapsibleContainer {
-  border: var(--basal);
-}
-
-.wrapper {
-  margin-bottom: 15px;
-}
